Huge rename `wiki` → `wiki_pages`
