+++ modules/wikis/app/components/wikis/wiki_pages/sub_header_component.rb
@@ -29,37 +29,41 @@
 #++
 
 module Wikis
-  class TableComponent < ::OpPrimer::BorderBoxTableComponent
-    columns :title, :project_name, :sub_pages_count, :last_edited
+  module WikiPages
+    class SubHeaderComponent < ApplicationComponent
+      alias_method :query, :model
 
-    mobile_columns :title, :project_name
-    mobile_labels :project_name
+      def initialize(query:)
+        super(query)
+      end
 
-    main_column :title
+      def filter_input_value
+        query.find_active_filter(:name)&.values&.first
+      end
 
-    def mobile_title
-      I18n.t("wikis.index.menu_title")
-    end
+      def sub_header_data_attributes
+        {
+          controller: "filter--filters-form",
+          "filter--filters-form-perform-turbo-requests-value": true,
+          "filter--filters-form-url-path-name-value": wiki_pages_path,
+          "filter--filters-form-output-format-value": "json",
+          "filter--filters-form-clear-button-id-value": clear_button_id,
+          test_selector: "wikis-sub-header"
+        }
+      end
 
-    def headers
-      [
-        [:title,           { caption: I18n.t("wikis.index.column_name") }],
-        [:project_name,    { caption: I18n.t("wikis.index.column_project") }],
-        [:sub_pages_count, { caption: I18n.t("wikis.index.column_sub_pages") }],
-        [:last_edited,     { caption: I18n.t("wikis.index.column_last_edited") }]
-      ]
-    end
+      def filter_input_data_attributes
+        {
+          "filter-name": "name",
+          "filter-type": "string",
+          "filter-operator": "~",
+          "filter--filters-form-target": "simpleFilter filterValueContainer simpleValue"
+        }
+      end
 
-    def blank_title
-      I18n.t("wikis.index.no_results_title")
-    end
-
-    def blank_description
-      I18n.t("wikis.index.no_results_description")
-    end
-
-    def blank_icon
-      :book
+      def clear_button_id
+        "wikis-filters-form-clear-button"
+      end
     end
   end
 end
