+++ modules/wikis/app/views/wikis/wiki_pages/index.html.erb
@@ -29,9 +29,9 @@ See COPYRIGHT and LICENSE files for more details.
 <% html_title t("wikis.index.menu_title") %>
 
 <% content_for :content_header do %>
-  <%= render(Wikis::IndexPageHeaderComponent.new(query: @query)) %>
+  <%= render(Wikis::WikiPages::IndexPageHeaderComponent.new(query: @query)) %>
 <% end %>
 
-<%= render(Wikis::SubHeaderComponent.new(query: @query)) %>
+<%= render(Wikis::WikiPages::SubHeaderComponent.new(query: @query)) %>
 
-<%= render(Wikis::IndexResultsComponent.new(pages: @pages)) %>
+<%= render(Wikis::WikiPages::IndexResultsComponent.new(wiki_pages: @wiki_pages)) %>
